refactor(legacy): archive speech lookup runtime
