+++ api_swedeb/core/load.py
@@ -1,12 +1,10 @@
-import abc
-import json
 import os
-import zipfile
 from os.path import isfile, join
 
 import pandas as pd
 from loguru import logger
 
+from api_swedeb.legacy.load import Loader, ZipLoader
 from penelope.corpus import VectorizedCorpus
 
 from .utility import time_call
@@ -142,46 +140,3 @@ def zero_fill_filename_sequence(name: str) -> str:
     return '-'.join(parts)
 
 
-class Loader(abc.ABC):
-    @abc.abstractmethod
-    def load(self, protocol_name: str) -> tuple[dict, list[dict]]: ...
-
-
-class ZipLoader(Loader):
-    def __init__(self, folder: str):
-        self.folder: str = folder
-
-    @staticmethod
-    def _resolve_payload_member(fp: zipfile.ZipFile, protocol_name: str) -> str:
-        payload_members = sorted(name for name in fp.namelist() if name.endswith(".json") and name != "metadata.json")
-        if len(payload_members) == 1:
-            return payload_members[0]
-
-        raise FileNotFoundError(f"JSON payload for {protocol_name} not found in archive")
-
-    def load(self, protocol_name: str) -> tuple[dict, list[dict]]:
-        """Loads tagged protocol data from archive"""
-        parts: list[str] = protocol_name.split('-')
-        sub_folder: str = parts[1]
-        candidate_files: list[str] = [
-            join(self.folder, sub_folder, f"{protocol_name}.zip"),
-            join(self.folder, f"{protocol_name}.zip"),
-            join(self.folder, '-'.join(parts[:-1] + [parts[-1].zfill(3)]) + ".zip"),
-            join(self.folder, '-'.join(parts[:-1] + [parts[-1].zfill(4)]) + ".zip"),
-            join(self.folder, '-'.join(parts[:-1] + [parts[-1].lstrip('0')]) + ".zip"),
-        ]
-        for filename in candidate_files:
-            if not os.path.isfile(filename):
-                continue
-            with zipfile.ZipFile(filename, "r") as fp:
-                metadata_str: bytes = fp.read("metadata.json")
-                metadata: dict = json.loads(metadata_str)
-                payload_member = self._resolve_payload_member(fp, protocol_name)
-                json_str: bytes = fp.read(payload_member)
-            # Use the archive stem as the canonical protocol key. This keeps
-            # zero-padded archive names stable without over-normalizing mixed
-            # suffixes like ``prot-1886--ak--040-01``.
-            metadata["name"] = os.path.splitext(os.path.basename(filename))[0]
-            utterances: list[dict] = json.loads(json_str)
-            return metadata, utterances
-        raise FileNotFoundError(protocol_name)
