+++ api_swedeb/legacy/__init__.py
@@ -1,0 +1,9 @@
+from .load import Loader, ZipLoader
+from .speech_lookup import SpeechTextRepository, SpeechTextService
+
+__all__ = [
+    "Loader",
+    "ZipLoader",
+    "SpeechTextRepository",
+    "SpeechTextService",
+]
