+++ tests/api_swedeb/core/test_speech_parity.py
@@ -39,7 +39,7 @@ from loguru import logger
 from api_swedeb.core import codecs as md
 from api_swedeb.core.configuration import ConfigStore, ConfigValue
 from api_swedeb.core.load import load_speech_index
-from api_swedeb.core.speech_text import SpeechTextRepository
+from api_swedeb.legacy.speech_lookup import SpeechTextRepository
 from api_swedeb.workflows.prebuilt_speech_index.build import SpeechCorpusBuilder
 
 # ---------------------------------------------------------------------------
